+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.5]:
+
+* Add support for Liquid-Bitcoin Master Blinding Key export 
+* Add support to Enable/disable optional features in Satochip v0.14-0.5+
+
 ## [0.15.4]:
 
 * Add support for Nostr event signature in satochip_cli CLI.